requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langgraph==0.2.34
langchain-openai==0.1.22
langchain-core==0.2.38
pysam==0.22.1
pydantic==2.7.4
python-dotenv==1.0.1
